Operations/backend/requirements.txt
Django>=5.2.0,<5.3.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.1.0,<4.2.0
Pillow>=10.0.0,<10.1.0
pytz>=2023.3
python-dateutil>=2.8.2,<2.9.0
drf-yasg>=1.21.5,<1.22.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
